// frontend/src/types/api.ts
// let's imagine this file is autogenerated from the backend
// ideally, we want to keep these api related types in sync
// with the backend instead of manually writing them out

import { TaskStatusEnum } from "@/enums/task-status.enum";

export type BaseEntity = {
  _id: string;
  createdAt: string;
  updatedAt: string;
};

export type Entity<T> = {
  [K in keyof T]: T[K];
} & BaseEntity;

export type Pagination = {
  currentPage: number;
  totalPages: number;
  totalItems: number;
  itemsPerPage: number;
  hasNextPage: boolean;
  hasPrevPage: boolean;
};

export type User = Entity<{
  email: string;
}>;

export type AuthResponse = {
  message: string;
  data: User
};


export type Task = Entity<{
  title: string;
  description: string;
  status: TaskStatusEnum;
  user: User;
}>;